pyproject.toml
[project]
name = "chatterbox-tts"
version = "0.2.0"
description = "Chatterbox: Open Source TTS and Voice Conversion by Resemble AI"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "resemble-ai", email = "engineering@resemble.ai"}
]
dependencies = [
    "numpy>=1.26.0",
    "librosa==0.11.0",
    "s3tokenizer",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "transformers>=4.46.3",
    "diffusers>=0.29.0",
    "conformer>=0.3.2",
    "safetensors>=0.5.3",
    "pyloudnorm",
    "ffmpeg-python",
    "jiwer",
    "nltk",
    "openai-whisper",
    "faster-whisper",
]

[project.urls]
Homepage = "https://github.com/resemble-ai/chatterbox"
Repository = "https://github.com/resemble-ai/chatterbox"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
